+++ src/Annotation/TypeScriptInterface.php
@@ -2,15 +2,14 @@
 
 namespace Paneon\PhpToTypeScript\Annotation;
 
-use Attribute;
-
 /**
  * Class TypeScriptInterface
  *
+ * @deprecated Use \Paneon\PhpToTypeScript\Attribute\TypeScript instead
  * @Annotation
  * @Target("CLASS")
  */
-#[Attribute(flags: Attribute::TARGET_CLASS)]
+#[\Attribute(flags: \Attribute::TARGET_CLASS)]
 class TypeScriptInterface
 {
     public $if = true;