+++ tests/Fixtures/AttributeClass.php
@@ -6,10 +6,10 @@ use DateTime;
 use FixtureInterface;
 use Paneon\PhpToTypeScript\Annotation\Exclude;
 use Paneon\PhpToTypeScript\Annotation\Type;
-use Paneon\PhpToTypeScript\Annotation\TypeScriptInterface;
 use Paneon\PhpToTypeScript\Annotation\VirtualProperty;
+use Paneon\PhpToTypeScript\Attribute\TypeScript;
 
-#[TypeScriptInterface]
+#[TypeScript]
 class AttributeClass
 {
     use SomeTrait;